Move cached meraki items to Meraki namespace
