+++ Pages/Meraki/Networks.cshtml.cs
@@ -2,6 +2,7 @@ using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.AspNetCore.Mvc.RazorPages;
 using Microsoft.EntityFrameworkCore;
+using QRStickers.Meraki;
 using System.Security.Claims;
 
 namespace QRStickers.Pages.Meraki;
