+++ Pages/Meraki/Organizations.cshtml.cs
@@ -1,6 +1,7 @@
 using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc.RazorPages;
 using Microsoft.EntityFrameworkCore;
+using QRStickers.Meraki;
 using System.Security.Claims;
 
 namespace QRStickers.Pages.Meraki;
